Renaming bookingReducer, display appointments
